services/script_generator/requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
firebase-admin==6.2.0
requests==2.31.0
gunicorn==21.2.0
python-jose==3.3.0
flask-cors==4.0.0
openai==1.12.0 